ReadNoFence returns INT32.
In RingBuffer we are comparing that to a UINT32.
The code casts the result to UINT64.
This is confusing.
Does it sign extend or not?
i.e. There is both a change in size and signedness. In which order?

The change in size is not required. Cast to UINT32 instead.

Related work items: #57987809, #60275790

+++ MsvmPkg/EmclDxe/RingBuffer.c
@@ -545,7 +545,7 @@ PkCompleteRemoval(
     // Mark that an interrupt is expected if the ring is now empty.
     //
 
-    if ((UINT64)ReadNoFence((UINT32*)&control->In) == NewOut)
+    if ((UINT32)ReadNoFence((UINT32*)&control->In) == NewOut)
     {
         PkpExpectInterrupt(PkLibContext, TRUE);
     }
